[Gold V] Title: 수 이어 쓰기 2, Time: 100 ms, Memory: 14124 KB -BaekjoonHub

+++ 백준/Gold/1790. 수 이어 쓰기 2/수 이어 쓰기 2.java
@@ -14,7 +14,7 @@ public class Main {
         }
         k -= 1;
 
-        long num = (int) Math.pow(10, len - 1) + k / len;
+        long num = (long) Math.pow(10, len - 1) + k / len;
         if (num > n) System.out.println(-1);
         else System.out.println(String.valueOf(num).charAt((int) (k % len)));
     }
